Display song duration and refresh at appropriate time.
Since we know how much time is left in the song, refresh the page when
that time has elapsed.

Don't want to display "time remaining" for the current song because
it'll be immediately out of date unless I implement some sort of JS
countdown which I don't want to do.